refactor(ui): namespaces
