+++ ui/src/components/namespaces/components/content/Dependencies.vue
@@ -53,8 +53,7 @@
     import dagre from "dagre";
     import ArrowExpandAll from "vue-material-design-icons/ArrowExpandAll.vue";
 
-    import {cssVariable} from "@kestra-io/ui-libs";
-    import {DependenciesNode} from "@kestra-io/ui-libs";
+    import {cssVariable, DependenciesNode} from "@kestra-io/ui-libs";
 
     import {linkedElements} from "../../../../utils/vueFlow";
     import {apiUrl} from "override/utils/route";
