+++ ui/src/components/namespaces/utils/useHelpers.ts
@@ -2,6 +2,13 @@ import {Component} from "vue";
 import {useRoute} from "vue-router";
 import {useI18n} from "vue-i18n";
 
+import BlueprintsBrowser from "../../../override/components/flows/blueprints/BlueprintsBrowser.vue";
+import Dashboard from "../../../components/dashboard/Dashboard.vue";
+import Flows from "../../../components/flows/Flows.vue";
+import Executions from "../../../components/executions/Executions.vue";
+import Dependencies from "../../../components/namespaces/components/content/Dependencies.vue";
+import EditorView from "../../../components/inputs/EditorView.vue";
+
 export interface Tab {
     locked?: boolean;
 
@@ -39,7 +46,7 @@ export function useHelpers() {
 
     const namespace = route.params?.id as string;
 
-    const parts = namespace.split(".") ?? [];
+    const parts = namespace?.split(".") ?? [];
     const details: Details = {
         title: parts.at(-1) || t("namespaces"),
         breadcrumb: [
@@ -58,5 +65,49 @@ export function useHelpers() {
         ],
     };
 
-    return {details};
+    const tabs: Tab[] = [
+        // If it's a system namespace, include the blueprints tab
+        ...(namespace === "system"
+            ? [
+                  {
+                      name: "blueprints",
+                      title: t("blueprints.title"),
+                      component: BlueprintsBrowser,
+                      props: {tab: "community", system: true},
+                  },
+              ]
+            : []),
+        {
+            name: "overview",
+            title: t("overview"),
+            component: Dashboard,
+            props: {containerClass: "full-container flex-0"},
+        },
+        {
+            name: "flows",
+            title: t("flows"),
+            component: Flows,
+            props: {topbar: false},
+        },
+        {
+            name: "executions",
+            title: t("executions"),
+            component: Executions,
+            props: {topbar: false, visibleCharts: true},
+        },
+        {
+            name: "dependencies",
+            title: t("dependencies"),
+            component: Dependencies,
+            props: {type: "dependencies", namespace},
+        },
+        {
+            name: "files",
+            title: t("files"),
+            component: EditorView,
+            props: {isNamespace: true, isReadOnly: false},
+        },
+    ];
+
+    return {details, tabs};
 }
