+++ ui/src/override/components/namespaces/useTabs.ts
@@ -1,18 +1,15 @@
 import {useRoute} from "vue-router";
 import {useI18n} from "vue-i18n";
 
-import {Tab, ORDER} from "../../../components/namespaces/utils/useHelpers";
+import {
+    Tab,
+    ORDER,
+    useHelpers,
+} from "../../../components/namespaces/utils/useHelpers";
 
 import DemoNamespace from "../../../components/demo/Namespace.vue";
 
-import BlueprintsBrowser from "../flows/blueprints/BlueprintsBrowser.vue";
-
-import Dashboard from "../../../components/dashboard/Dashboard.vue";
-import Flows from "../../../components/flows/Flows.vue";
-import Executions from "../../../components/executions/Executions.vue";
-import Dependencies from "../../../components/namespaces/components/content/Dependencies.vue";
 import KVTable from "../../../components/kv/KVTable.vue";
-import EditorView from "../../../components/inputs/EditorView.vue";
 
 const lockedProps = (tab: string) => ({
     locked: true,
@@ -27,45 +24,11 @@ export function useTabs() {
     const namespace = route.params?.id as string;
 
     const tabs: Tab[] = [
-        // If it's a system namespace, include the blueprints tab
-        ...(namespace === "system"
-            ? [
-                  {
-                      name: "blueprints",
-                      title: t("blueprints.title"),
-                      component: BlueprintsBrowser,
-                      props: {tab: "community", system: true},
-                  },
-              ]
-            : []),
-        {
-            name: "overview",
-            title: t("overview"),
-            component: Dashboard,
-            props: {containerClass: "full-container flex-0"},
-        },
+        ...useHelpers().tabs,
         {
             ...lockedProps("edit"),
             name: "edit",
             title: t("edit"),
-        },
-        {
-            name: "flows",
-            title: t("flows"),
-            component: Flows,
-            props: {topbar: false},
-        },
-        {
-            name: "executions",
-            title: t("executions"),
-            component: Executions,
-            props: {topbar: false, visibleCharts: true},
-        },
-        {
-            name: "dependencies",
-            title: t("dependencies"),
-            component: Dependencies,
-            props: {type: "dependencies", namespace},
         },
         {
             ...lockedProps("secrets"),
@@ -87,12 +50,6 @@ export function useTabs() {
             title: t("kv.name"),
             component: KVTable,
             props: {namespace},
-        },
-        {
-            name: "files",
-            title: t("files"),
-            component: EditorView,
-            props: {isNamespace: true, isReadOnly: false},
         },
         {
             ...lockedProps("history"),
